Rationalizing the Interface class
The constructor, the __init__ method, now takes a bunch of arguments.
It doesn't care where the arguments come from - a file, from discovery,
from the CLI.
The discover method returns a dictionary of Interfaces, key'd by
interface name.
If the system configuration comes from a file, then the method that
handles the file may call the constructor directly.