Rationalize the use of the variables current_system and nominal_system.
In general, I want to look at the system as it is.  At boot time, I
want to see if I have a configuration that works.  When I am testing,
I want to see if I have a system that works.  When I am diagnosing,
I want to see if there are any changes from a nominal system.  One
of those changes broke the network.  When I am monitoring, I want to
see if there is any change.