Issue 27 and Issue 28
For issue 27, it was not clear that the filename after the --diagnose
switch is optional.  It was also not clear that the filename was a
configuration filename.

For issue 28, testing for the string "nominal " in syserr_str is
failing.  But there is a bunch of other stuff in that variable.
Try running the command standalone instead of spawning a subprocess