Working on Issue 29
There is a circular import problem with class SystemDescription.  So I
moved the definition of class SystemDescription out of nbmdt and into
utilities.py, which is probably where it ought to be anyway. However,
now OsCliInter is having an import proleml.  I'd solve it, but I
want to go to bed now.